#1356 - Rename lancaster module to smile
- Renamed module, artifactId and packages
- Also renamed class to SmileLancasterStemmer (there is also a Porter stemmer which is not integrated yet)
