+++ dkpro-core-asl/pom.xml
@@ -498,7 +498,7 @@
       </dependency>
       <dependency>
         <groupId>org.dkpro.core</groupId>
-        <artifactId>org-dkpro-core-lancaster-asl</artifactId>
+        <artifactId>org-dkpro-core-smile-asl</artifactId>
         <version>2.0.0-SNAPSHOT</version>
       </dependency>
       <dependency>
@@ -640,7 +640,7 @@
     <module>../dkpro-core-readability-asl</module>
     <module>../dkpro-core-rftagger-asl</module>
     <module>../dkpro-core-snowball-asl</module>
-    <module>../dkpro-core-lancaster-asl</module>
+    <module>../dkpro-core-smile-asl</module>
     <module>../dkpro-core-stopwordremover-asl</module>
     <module>../dkpro-core-testing-asl</module>
     <module>../dkpro-core-textcat-asl</module>
