+++ dkpro-core-smile-asl/LICENSE.txt
