+++ dkpro-core-smile-asl/pom.xml
@@ -25,9 +25,9 @@
     <version>2.0.0-SNAPSHOT</version>
     <relativePath>../dkpro-core-asl</relativePath>
   </parent>
-  <artifactId>dkpro-core-lancaster-asl</artifactId>
+  <artifactId>dkpro-core-smile-asl</artifactId>
   <packaging>jar</packaging>
-  <name>DKPro Core ASL - Lancaster (v ${smile.version}) (ASL</name>
+  <name>DKPro Core ASL - Smile (v ${smile.version}) (ASL</name>
   <url>https://dkpro.github.io/dkpro-core/</url>
   <description>http://haifengl.github.io/smile</description>
   <properties>
