+++ dkpro-core-smile-asl/src/test/resources/Lancaster_test_rules.txt
