+++ dkpro-core-smile-asl/src/test/resources/log4j2.xml
